more improvement

+++ src/Controllers/SubscriberController.php
@@ -19,11 +19,6 @@ class SubscriberController
         $this->validator = new ValidationService();
     }
 
-    public function setTestSubscriberModel($model)
-    {
-        $this->subscriberModel = $model;
-    }
-
     public function getSubscriber(string $phoneNumber)
     {
         if ($this->subscriberModel == null) {
@@ -43,36 +38,55 @@ class SubscriberController
         }
     }
 
-    public function addSubscriber()
+    public function addSubscriber(array $newSubscriber = null)
     {
-        $newSubscriber = json_decode(file_get_contents('php://input'), true);
+        $newSubscriber = $newSubscriber ?? $this->getRequestedData();
 
         if (!$this->validator->validateSubscriberData($newSubscriber)) {
             http_response_code($this->validator->getResponseStatus());
             echo $this->validator->getResponse();
+            return;
         }
 
         try {
-            $id = $this->model->addSubscriber($newSubscriber);
+            if ($this->subscriberModel == null) {
+                $this->subscriberModel = new SubscriberModel(new Database());
+            }
+            $id = $this->subscriberModel->addSubscriber($newSubscriber);
             http_response_code(201); // Created
-            echo json_encode(['message' => 'Subscriber added']);
+            echo json_encode([
+                'message' => "Subscriber added ID: {$id}, Phone Number: {$newSubscriber['phoneNumber']}",
+                "id" => $id,
+                "phoneNumber" => $newSubscriber['phoneNumber']
+            ]);
         } catch (PDOException $e) {
             http_response_code(500);
             echo json_encode(['message' => 'Failed to add subscriber']);
         }
     }
 
-    public function updateSubscriber()
+    public function updateSubscriber(array $updatedSubscriber = null)
     {
-        $updatedSubscriber = json_decode(file_get_contents('php://input'), true);
+        $updatedSubscriber = $updatedSubscriber ?? $this->getRequestedData();
 
         if (!$this->validator->validateSubscriberData($updatedSubscriber)) {
             http_response_code($this->validator->getResponseStatus());
             echo $this->validator->getResponse();
+            return;
         }
 
         try {
-            $updated = $this->model->updateSubscriber($updatedSubscriber);
+            if ($this->subscriberModel == null) {
+                $this->subscriberModel = new SubscriberModel(new Database());
+            }
+
+            if (!$this->subscriberModel->phoneNumberExists($updatedSubscriber['phoneNumber'])) {
+                http_response_code(404);
+                echo json_encode(['message' => 'Subscriber not found']);
+                return;
+            }
+
+            $updated = $this->subscriberModel->updateSubscriber($updatedSubscriber);
             if ($updated) {
                 http_response_code(200);
                 echo json_encode(['message' => 'Subscriber updated']);
@@ -89,10 +103,19 @@ class SubscriberController
     public function deleteSubscriber(string $phoneNumber)
     {
         try {
-            $deleted = $this->model->deleteSubscriber($phoneNumber);
+            if ($this->subscriberModel == null) {
+                $this->subscriberModel = new SubscriberModel(new Database());
+            }
+
+            if (!$this->subscriberModel->phoneNumberExists($phoneNumber)) {
+                http_response_code(404);
+                echo json_encode(['message' => 'Subscriber not found']);
+                return;
+            }
+            $deleted = $this->subscriberModel->deleteSubscriber($phoneNumber);
             if ($deleted) {
                 http_response_code(200);
-                echo json_encode(['message' => 'Subscriber deleted']);
+                echo json_encode(['message' => "Subscriber {$phoneNumber} deleted"]);
             } else {
                 http_response_code(404);
                 echo json_encode(['message' => 'Subscriber not found']);
@@ -106,24 +129,56 @@ class SubscriberController
     public function getAllSubscribers(int $limit = 100, int $offset = 0)
     {
         try {
+            if ($this->subscriberModel == null) {
+                $this->subscriberModel = new SubscriberModel(new Database());
+            }
+
             if ($limit > 100) {
-                $this->model->setLimit($limit);
+                $this->subscriberModel->setLimit($limit);
             }
             if ($offset > 0) {
-                $this->model->setOffset($offset);
+                if ($offset > $limit) {
+                    $this->subscriberModel->setOffset($offset);
+                } else {
+                    $this->subscriberModel->setOffset(0);
+                }
             }
-            $subscribers = $this->model->getAllSubscribers();
+
+            $subscribers = $this->subscriberModel->getAllSubscribers();
+
+            // Calculate the next and previous offsets
+            $nextOffset = $offset + $limit;
+            $prevOffset = $offset - $limit;
+
+            // Ensure previous offset doesn't go below 0
+            $prevOffset = $prevOffset < 0 ? 0 : $prevOffset;
 
             foreach ($subscribers as &$resource) {
                 $resource['password'] = ''; // Masking the password
                 $resource["features"] = isset($resource["features"]) ? json_decode($resource["features"], true) : [];
+
+                // Build the pagination links
+                $resource["links"] = [
+                    'prev' => "/ims/subscriber/all/{$limit}/{$prevOffset}",
+                    'next' => "/ims/subscriber/all/{$limit}/{$nextOffset}"
+                ];
             }
 
             http_response_code(200);
             echo json_encode($subscribers);
         } catch (\Throwable $th) {
             http_response_code(500);
-            echo json_encode(['message' => 'Failed to retrieve subscribers'] . $th->getMessage());
+            echo json_encode(['message' => 'Failed to retrieve subscribers']);
         }
+    }
+
+    public function setTestSubscriberModel($model)
+    {
+        $this->subscriberModel = $model;
+    }
+
+    public function getRequestedData()
+    {
+        return json_decode(file_get_contents('php://input'), true);
     }
 }
